2026-03-10-chainlit/demo/advanced/requirements.txt
chainlit>=2.10.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0
python-dotenv>=1.0.0
tavily-python>=0.5.0
requests>=2.31.0
aiohttp>=3.9.0
pydantic>=2.0.0
